Creating saving functions
The program contains functions for copying, saving and loading class instances. Either an entire simulation or a specified class instance like a particular flower, environment, bee or swarm. When loading only the reintroduction of a simulation is currently supported and other code might be needed to allow it even if the object is loaded correctly